+++ src/shared/mocks/mockChatData.ts
@@ -3,7 +3,7 @@ export const mockChatData = [
     id: '1',
     type: 'group',
     participants: 3,
-    groupName: '호두',
+    title: '호두',
     latestMessage: '티켓 예매는 제가 한 번에 할까요?',
     timestamp: '방금',
     unreadCount: 4,
@@ -11,10 +11,10 @@ export const mockChatData = [
   {
     id: '2',
     type: 'single',
-    groupName: '야구조아',
+    title: '야구조아',
     participants: 1,
     latestMessage: '그럼 저희 경기 당일에 매표소 앞에서 뵐...',
     timestamp: '어제',
     unreadCount: 0,
   },
-];
+] as const;
